Fix PowerUser type in Exercise 7
I swear, this used to work before! But I checked this project out
today and the "role" and "occupation" properties for the PowerUser
object in the Persons array started throwing this error:

Type 'string' is not assignable to type 'never'.

For some reason, I now have to omit the "type" property first before
intersecting, rather than after.